+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Fix
 - [#755 - [BUG] [TRACE] [RS core Monitoring] Field "duplicate" not set to "true" when several processing are linked to same chunk or DSIB](https://github.com/COPRS/rs-issues/issues/755)
 - Use OneToOne relation instead of ManyToOne between Product and InputListInternal/OutputList
+- Extract leaf for each level of the tree and not only the first one
 
 ## [1.4.0-rc1] - 2022-11-23
 ### Added